Fixed file uploadind and deleting error

+++ app/Http/Controllers/LandlordListingController.php
@@ -157,7 +157,9 @@ class LandlordListingController extends Controller
 
     private function _changeFileName(Request $request, string $filename)
     {
-        $newFileName = $filename.'-'.$request->user()->first_name.'-'.$request->user()->last_name;
+        $file = \substr($filename, 0, -4);
+        $fileExtension = \substr($filename, -4);
+        $newFileName = $file.'-'.$request->user()->first_name.'-'.$request->user()->last_name.$fileExtension;
         return $newFileName;
     }
 
@@ -197,42 +199,42 @@ class LandlordListingController extends Controller
         // Format and store file to storage
         $gasCertificate = $this->_changeFileName($request, $request->file('gas_certificate')->getClientOriginalName());
         $listingDocument->gas_certificate = $gasCertificate;
-        $request->file('gas_certificate')->storeAs('listing/documents', $gasCertificate);
+        $request->file('gas_certificate')->storeAs('public/listing/documents', $gasCertificate);
         $listingDocument->gas_certificate_expiry_date = $request->gas_certificate_expiry_date;
 
         $electricalCertificate = $this->_changeFileName($request, $request->file('electrical_certificate')->getClientOriginalName());
         $listingDocument->electrical_certificate = $electricalCertificate;
-        $request->file('electrical_certificate')->storeAs('listing/documents', $electricalCertificate);
+        $request->file('electrical_certificate')->storeAs('public/listing/documents', $electricalCertificate);
         $listingDocument->electrical_certificate_expiry_date = $request->electrical_certificate_expiry_date;
 
         $detectorsCertificate = $this->_changeFileName($request, $request->file('detectors_certificate')->getClientOriginalName());
         $listingDocument->detectors_certificate = $detectorsCertificate;
-        $request->file('detectors_certificate')->storeAs('listing/documents', $detectorsCertificate);
+        $request->file('detectors_certificate')->storeAs('public/listing/documents', $detectorsCertificate);
         $listingDocument->detectors_certificate_expiry_date = $request->detectors_certificate_expiry_date;
 
         $emergency_lightingCertificate = $this->_changeFileName($request, $request->file('emergency_lighting_certificate')->getClientOriginalName());
         $listingDocument->emergency_lighting_certificate = $emergency_lightingCertificate;
-        $request->file('emergency_lighting_certificate')->storeAs('listing/documents', $emergency_lightingCertificate);
+        $request->file('emergency_lighting_certificate')->storeAs('public/listing/documents', $emergency_lightingCertificate);
         $listingDocument->emergency_lighting_certificate_expiry_date = $request->emergency_lighting_certificate_expiry_date;
 
         $fire_riskCertificate = $this->_changeFileName($request, $request->file('fire_risk_certificate')->getClientOriginalName());
         $listingDocument->fire_risk_certificate = $fire_riskCertificate;
-        $request->file('fire_risk_certificate')->storeAs('listing/documents', $fire_riskCertificate);
+        $request->file('fire_risk_certificate')->storeAs('public/listing/documents', $fire_riskCertificate);
         $listingDocument->fire_risk_certificate_expiry_date = $request->fire_risk_certificate_expiry_date;
 
         $patCertificate = $this->_changeFileName($request, $request->file('pat_certificate')->getClientOriginalName());
         $listingDocument->pat_certificate = $patCertificate;
-        $request->file('pat_certificate')->storeAs('listing/documents', $patCertificate);
+        $request->file('pat_certificate')->storeAs('public/listing/documents', $patCertificate);
         $listingDocument->pat_certificate_expiry_date = $request->pat_certificate_expiry_date;
 
         $insuranceCertificate = $this->_changeFileName($request, $request->file('insurance_certificate')->getClientOriginalName());
         $listingDocument->insurance_certificate = $insuranceCertificate;
-        $request->file('insurance_certificate')->storeAs('listing/documents', $insuranceCertificate);
+        $request->file('insurance_certificate')->storeAs('public/listing/documents', $insuranceCertificate);
         $listingDocument->insurance_certificate_expiry_date = $request->insurance_certificate_expiry_date;
 
         $ownershipCertificate = $this->_changeFileName($request, $request->file('ownership_certificate')->getClientOriginalName());
         $listingDocument->ownership_certificate = $ownershipCertificate;
-        $request->file('ownership_certificate')->storeAs('listing/documents', $ownershipCertificate);
+        $request->file('ownership_certificate')->storeAs('public/listing/documents', $ownershipCertificate);
         $listingDocument->ownership_certificate_expiry_date = $request->ownership_certificate_expiry_date;
 
         if ($request->has('proof')) {
@@ -264,17 +266,20 @@ class LandlordListingController extends Controller
         $listingImage = new ListingImage;
 
         if ($request->file()) {
-            $filename = $request->file->getClientOriginalName().'-'.time();
-            $filepath = $request->file('file')->storeAs('listing/images', $filename);
+            $filenameWithExt = $request->file->getClientOriginalName();
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            $fileExtension = $request->file->getClientOriginalExtension();
+            $filenameToStore = $filename.'-'.time().'-'.$fileExtension;
+            $filepath = $request->file('file')->storeAs('public/listing/images', $filenameToStore);
 
             $listingImage->listing_id = $request->listing_id;
-            $listingImage->image_name = $filename;
+            $listingImage->image_name = $filenameToStore;
             $listingImage->save();
 
             return redirect()->route('landlord.index')->with('success', 'Property was added successfully');
         }
 
-        return \redirect()->route('listing.add.submit_images')->withErrors('An error occurred. Please try again.');
+        return redirect()->route('listing.add.submit_images')->withErrors('An error occurred. Please try again.');
         
     }
 
@@ -282,17 +287,27 @@ class LandlordListingController extends Controller
     {
         $listingImages = ListingImage::where('listing_id', '=', $id)->get();
         foreach ($listingImages as $image) {
-            Storage::delete('/storage/listing/images/', $image);
+            unlink(public_path('storage/listing/images/'.$image->image_name));
         }
-
-        $listingDocuments = ListingDocuments::where('listing_id', '=', $id);
+            
+        $listingDocuments = ListingDocuments::where('listing_id', '=', $id)->get();
         foreach ($listingDocuments as $document) {
-            Storage::delete('/storage/listing/documents', $document);
+            unlink(public_path('storage/listing/documents/'.$document->gas_certificate));
+            unlink(public_path('storage/listing/documents/'.$document->electrical_certificate));
+            unlink(public_path('storage/listing/documents/'.$document->detectors_certificate));
+            unlink(public_path('storage/listing/documents/'.$document->emergency_lighting_certificate));
+            unlink(public_path('storage/listing/documents/'.$document->fire_risk_certificate));
+            unlink(public_path('storage/listing/documents/'.$document->pat_certificate));
+            unlink(public_path('storage/listing/documents/'.$document->insurance_certificate));
+            unlink(public_path('storage/listing/documents/'.$document->ownership_certificate));
         }
-
-        if(ClientGroup::where('listing_id', '=', $id)->delete()) {
-            if(Listing::destroy($id)) {
-                return redirect()->route('listing.index')->with('success', 'Listing has been deleted successfully');
+        if (ListingImage::where('listing_id', '=', $id)->delete()) {
+            if (ListingDocuments::where('listing_id', '=', $id)->delete()) {
+                if(ClientGroup::where('listing_id', '=', $id)->delete()) {
+                    if(Listing::destroy($id)) {
+                        return redirect()->route('listing.index')->with('success', 'Listing has been deleted successfully');
+                    }
+                }
             }
         }
